+++ src/main/java/com/atwoz/member/ui/auth/interceptor/LoginValidCheckerInterceptor.java
@@ -2,8 +2,8 @@ package com.atwoz.member.ui.auth.interceptor;
 
 import com.atwoz.member.domain.auth.TokenProvider;
 import com.atwoz.member.exception.exceptions.auth.LoginInvalidException;
-import com.atwoz.member.ui.auth.support.AuthenticationContext;
-import com.atwoz.member.ui.auth.support.AuthenticationExtractor;
+import com.atwoz.member.ui.auth.support.auth.AuthenticationContext;
+import com.atwoz.member.ui.auth.support.auth.AuthenticationExtractor;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
