+++ src/test/java/com/atwoz/helper/MockBeanInjection.java
@@ -2,7 +2,7 @@ package com.atwoz.helper;
 
 import com.atwoz.member.application.auth.AuthService;
 import com.atwoz.member.domain.auth.TokenProvider;
-import com.atwoz.member.ui.auth.support.AuthenticationContext;
+import com.atwoz.member.ui.auth.support.auth.AuthenticationContext;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.data.jpa.mapping.JpaMetamodelMappingContext;
