+++ src/test/java/com/atwoz/member/ui/auth/support/AuthenticationContextTest.java
@@ -1,6 +1,7 @@
 package com.atwoz.member.ui.auth.support;
 
 import com.atwoz.member.exception.exceptions.auth.LoginInvalidException;
+import com.atwoz.member.ui.auth.support.auth.AuthenticationContext;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
