+++ src/test/java/com/atwoz/member/ui/auth/support/AuthenticationExtractorTest.java
@@ -1,12 +1,12 @@
 package com.atwoz.member.ui.auth.support;
 
+import com.atwoz.member.ui.auth.support.auth.AuthenticationExtractor;
 import jakarta.servlet.http.HttpServletRequest;
+import java.util.Optional;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
-
-import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.SoftAssertions.assertSoftly;
